Fixing Elliptical Arc Close Bounds to Work Consistently
Reworked the calculations for finding the tight bounding box of a
rotated elliptical arc, so it will find the correct bounds no mater the
angles used.
Fixing The Contains method for figuring out if a point is contained
within an elliptical arc section, so it would return the correct results
when the ellipse is rotated.
Added another Circle Interpolator so there is now one for radians and 0
to 1.
Changed the Needle control to match the angles found in GDI.