Visual Studio Upgrade
Updating to latest release candidate of Visual Studio 2017.
Fixing capitalization errors.
Failed to get a .NET Standard library working with the winforms project.
I managed to get it to compile, but it crashes on initialization of the
form when the library is included on the winforms project.
Ran some autofixups provided by Rosylin in c#7.
Renamed the Paths folder from Figure.
Changed OnDeserializ(ed/ing) methods to private per code analysis.
Removed the [Pure] Attribute, as it is dubious what it actually does.