Prerequsites for Bug Fixes
Updated to latest release candidate of Visual Studio 2017.
Found and implemented a potential fix for issue #6 using a Polynomial
struct. The fix is not yet attached to the shape class.
Found and implemented a fix for #7 and #8 . The fix is part of the fix
for #6 and work is needed to get it to work with the older classes.
Prerequsite work started on fixing issue #3, with finding intersections
of component shapes. The work should make fixing #4 #5 and #2 possible.
Experemented with Intersection Locus idea. Doesn't seem all that
practical. Might stick with Point array and enum setup.
Deleted the Engine.Math library stub. Attampts to share a Universal, or
standard library aren't ready yet. I just haven't been able to figure
out how to get it to work yet.
Found ArrayUtilities wasn't working as expected. Made changes to
expectations.
Needed a Node revealer class to be able to see intersections.
Fixed some spelling errors.
Moved Line and Ray back up to Engine. Going to need them to test
intersections, once they are fully built out.
Added screen point class to enable adding a point to the vector map.
Added new Math constants used in the intersections tests.
Moved some experemental classes back to the Speed tester app.