Fixing Rectangle Intersects Rectangle
This fixes #14 to correct two issues with the Rectangle intersects
Rectangle method.
Issue 1. Was that a width was being used to calculate a y intersection
rather than a height as it should have.
Issue 2. Was that the height and width parameters were in the wrong
order in the parameter list. X or width should have gone before Y or the
height.
Also making GeometryPath IEnumerable, and cleaning up Connector and
PointChain a little.