Refactering Editor Tests and More
Moving Testing shapes out of the comments of the Editor form to a
seporate class with static methods to better organize the test cases.
Fixing namespace collisions with the starter seed code for the Direct2D
control.
A friend noticed some multiple initialization of array fields, so I went
through and cleaned up as many cases as I could find.
Added test properties to EllipticalArc class. They need to be changed to
caching properties.