Work on Polycurves
Renamed PathContour again, this time to PolycurveContour. This seems
more fitting for what it is.
Added a non-working Polycurve class to contain the PolycurveContour
objects. It is based off of the Polygon class.
Moved some documents around, to try to put them somewhere more logical.
Cleaned up MethodSpeedTester a little.