Infinate Lines Working, Fixed Intersections
Fixed up the Lines class and got it to render on screen.
Worked on Ray class. But Rays will look like Lines until a Ray Rectangle
method is implemented.
The Main Render method now requires a Bounds rectangle for Infinate
shapes.
Added a Generators class for objects that can be reduced to be
implemented as Polygons, Polycurves, or PolyBeziers.
Changed the IntersectionState enum to be a flag type, so there can be
combined intersection types. Still needs some thought to make it more
meaningfull.
Fixed the AppendPoints method on the Intersection struct so it will
acctualy add points.
Fixed up some broken implementations of Intersections methods in the
Intersections class.
Added Line Intersections to some of the shapes.
Made implementations of the intersections overloads methods private to
try to quash some potential bugs.