Intersections between Elliptical Arcs and Lines Fix
Added Trigonometric Identities to the Primer to help work through broken
Ellipse Intersection issue.
Changed the signatures of the distortion methods to be uniform.
Found that the Rotate method in the Distortions class was messed up.
This commit fixes #22 by translating the point to the origen, then
rotating and translating the point back once done.
Started an idea that the Matrix classes could possibly be IEnumerable.
Discovered that the intersection methods for Cubic Bezier curves and
lines/line segments does not work as expected. Tried reporting
intersection code to see if I could get the new port working, but it is
not. Did find that the oldest intersection code works for the cases the
new does not. Weird. Need to keep working on it.
Started adding intersection methods between shapes and Lines, in the
process discovered that Elliptical Arcs weren't working when the ellipse
was rotated. This commit fixes #21 by rewriting the code that checks
which side of the chord a point is on for the intersection bassed on the
code for interpolating an elliptical arc.