Fixing Line Cubic Bezier Intersections
Rewrote the intersection betweent Line/Line Segment and Quadratic/Cubic
Beziers. They are much more simple, and now produce better results.
Found and fixed a bug where the intersection methods between Lines/Line
Segments and Cubic Bezier curves would not produce results if the start
and end points, the intersecting line and the control nodes are all
parallel. The fix was to detect the parallelity using determinants of
the paralell features, then find the roots of the lower degree
polynomial.
Swapped out some of the Roots methods. The previous ones resulted in
incorrect results. Moved the Roots methods to the Maths class in
Maths.Algebraic.cs, and linked to them from the Polynomial class.
Boved BezierCoefficients to Maths.Algebraic.cs and created overloaded
methods for Quadratic and Linear Bezier curves.
Simplified StartAngle and SweepAngle in the ArcSegment class. The inline
methods were broken.
Renamed EllipsePolarAngle to EllipticalPolarAngle.