Updating to C# 7.1
Updated Solution to the latest preview build of Visual Studio.
Updated projects to use the "latest" version of C#. Now using c# 7.1.
Applied the Scale distort to the Quadratic Bezier to Quadratic Bezier
Intersection test case, to try to come up with a more simple way to
write Unit Test cases.
Changed the Filters implementation to use typeparams to see if it might
reduce the need to cast preserving filtered shapes back to the former
type.