Misc. Changes
Adding more comment headers.
Moved some classes to Engine.Experimental from the Method tester.
Changed the IFormatable ToString implementation in various
classes/structs so it is public.
Changed Bezier to Bézier in text and comments for spell check.
Renamed Maths.Right back to Maths.HalfPi.
Added readonly to a bunch of fields.
Fixed build dependencies.
Added accessibility modifiers for items that were missing them.
Continuing to try to fix polycurve point hit test.
Added an IPrimitive interface as the base of point and so on, to try out
an idea for how to redo polycurves.
Commented out DebuggerDispaly for some types. It seems they may not work
wellwith properties.
Changed AccumulatorPoint to a struct.