Working out Possible Refactoring
Working out a refactoring to split the big library into smaller specialized ones. Not all types have been fully recreated in the experiment. Mostly the project is up enough to get it to compile.
Also testing to see if it might be possible to use structs for shapes rather than classes to try to work Point2D in as a struct shape, and shape component. Also would let Vector2D be used to represent Vector fields.
Updated the Experiment to .NET 5 preview.