Switching to .NET 5.0
Replacing with work done on a copy to experiment with porting to .NET core.
Much of the Designer still needs work. but the forms open now.
Updating primitive structs to work split up into more focused libraries.
Did a little work on Midi project. Still needs a lot of work.
Figured out a way to split media files in different dlls and register them with the main library.
Cleaned some of the code.