Fixing Midi Parser
Fixing the midi parser so it will acctually load midi files into memory.
Changed the Expandable collections property descripters to pad the numbers with 0s on the left so they always sort in the same order in the property grid.
Deleting the ASCIIEncoding class as it is nolonger needed, seeing as .NET Core has what is needed to read ASCII.
Cleaning up some code.
Started modifying the coding conventions.